requirements.txt
streamlit>=1.32.0
groq>=0.11.0
python-dotenv>=1.0.0
pydub>=0.25.1
